pyproject.toml
[project]
name = "adk-weather-agent"
version = "0.1.0"
description = "Weather search agent built with Google Agent Development Kit"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.26.0",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "ruff>=0.15.5",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "A",   # flake8-builtins
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
    "RUF", # ruff-specific rules
]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = ["google.*"]
ignore_missing_imports = true
